Demos/Paint Color Recommendation using GenAI/requirements.txt
streamlit
# transformers
# tensorflow
google-cloud-aiplatform
langchain
google-cloud-discoveryengine